holiday_planner_gradio/requirements.txt
gradio>=4.36.0,<5
python-dotenv>=1.0.1